Use long int for LandmarkId, int would be too small if 16bits

+++ src/UtilsOpenCV.h
@@ -84,7 +84,7 @@ enum Kstatus {
 
 // Definitions relevant to frame types
 using FrameId = uint64_t; // Frame id is used as the index of gtsam symbol (not as a gtsam key).
-using LandmarkId = int; // -1 for invalid landmarks.
+using LandmarkId = long int; // -1 for invalid landmarks. // int would be too small if it is 16 bits!
 using LandmarkIds = std::vector<LandmarkId>;
 enum class LandmarkType {SMART, PROJECTION};
 using KeypointCV = cv::Point2f;
